Fixed initial stats bug
When the program is started, spurious stats were shown for the 'currently selected item' (as no item is selected and there's no way of detecting the fact). This fix solves that problem at the expense of not displaying the top item at all when it's the first one requested. In practice this isn't a significant problem.